restructure to pull in tools from module automatically
